requirements.txt
Werkzeug==2.0.3
Flask==2.0.1
flask_sqlalchemy==2.5.1
mysql-connector-python==8.0.28
